azure-func/requirements.txt
requests
azure-functions
psycopg2-binary
protobuf